Divide iroha-util (#51)
* feat: iroha-util devided into several files

* fix: remove default export from files

* feat: changed path to iroha-util

* feat: removed iroha-util
