+++ src/util/iroha/index.js
@@ -1,0 +1,9 @@
+import * as queries from './queries'
+import * as commands from './commands'
+import * as functions from './functions'
+
+export default {
+  ...queries,
+  ...commands,
+  ...functions
+}
