+++ src/util/iroha/util.js
@@ -1,0 +1,38 @@
+import Debug from 'debug'
+const debug = Debug('iroha-util')
+
+export const DEFAULT_TIMEOUT_LIMIT = 5000
+
+/**
+ * cached items available from start to end of the app
+ * plus, `nodeIp` is persisted by localStorage
+ */
+export const cache = {
+  username: null, // NOT persisted by localStorage
+  // TODO: do not cache keys; require a private key on every action instead.
+  // For now it is needed for queries until tokens are implemented.
+  key: null, // NOT persisted by localStorage
+  nodeIp: null // persisted by localStorage
+}
+
+const protoEnumName = {}
+export function getProtoEnumName (obj, key, value) {
+  if (protoEnumName.hasOwnProperty(key)) {
+    if (protoEnumName[key].length < value) {
+      return 'unknown'
+    } else {
+      return protoEnumName[key][value]
+    }
+  } else {
+    protoEnumName[key] = []
+    for (let k in obj) {
+      let idx = obj[k]
+      if (isNaN(idx)) {
+        debug(`getProtoEnumName:wrong enum value, now is type of ${typeof idx} should be integer`)
+      } else {
+        protoEnumName[key][idx] = k
+      }
+    }
+    return getProtoEnumName(obj, key, value)
+  }
+}
